+++ PlantUMLCodeGeneratorGUI/Processor.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Text.RegularExpressions;
+using PlantUMLCodeGeneratorGUI.classes;
 
 namespace PlantUMLCodeGeneratorGUI
 {
