+++ PlantUMLCodeGeneratorGUI/classes/Class.cs
@@ -4,13 +4,13 @@ using System.Linq;
 using System.Text.RegularExpressions;
 using PlantUMLCodeGeneratorGUI.classes.utils;
 
-namespace PlantUMLCodeGeneratorGUI
+namespace PlantUMLCodeGeneratorGUI.classes
 {
     class Class
     {
         public Namespace ParentNamespace;
         public string Name { get; internal set; }
-        public string FullName { get { return ParentNamespace.FullName + "::" + Name; } }
+        public string FullName => ParentNamespace.FullName + "::" + Name;
 
         public Scope PublicScope { get; internal set; }
         public Scope ProtectedScope { get; internal set; }
@@ -30,22 +30,21 @@ namespace PlantUMLCodeGeneratorGUI
 
         public static Class GetClass(Namespace namespaceObj, string className, bool justCheck = false)
         {
-            Class foundClass = null;
-            if (namespaceObj.Classes.TryGetValue(className, out foundClass) == false)
+            Class foundClass;
+            if (namespaceObj.Classes.TryGetValue(className, out foundClass)) return foundClass;
+
+            if (justCheck)
             {
-                if (justCheck)
+                foreach (var childNamespace in namespaceObj.ChildNamespaces)
                 {
-                    foreach (var childNamespace in namespaceObj.ChildNamespaces)
-                    {
-                        foundClass = GetClass(childNamespace, className, true);
-                        if (foundClass != null) break;
-                    }
+                    foundClass = GetClass(childNamespace, className, true);
+                    if (foundClass != null) break;
                 }
-                else
-                {
-                    foundClass = new Class(namespaceObj, className);
-                    namespaceObj.Classes.Add(foundClass.Name, foundClass);
-                }
+            }
+            else
+            {
+                foundClass = new Class(namespaceObj, className);
+                namespaceObj.Classes.Add(foundClass.Name, foundClass);
             }
 
             return foundClass;
@@ -255,7 +254,7 @@ namespace PlantUMLCodeGeneratorGUI
 
                             if (argument.Contains("=")) argument = argument.Substring(0, argument.IndexOf("=", StringComparison.Ordinal)).Trim();
 
-                            var hasASpace    = argument.Contains(' ');
+                            var hasASpace = argument.Contains(' ');
                             var argumentName = hasASpace ? argument.Split(' ').Last() : "";
                             var argumentType = hasASpace ? argument.Substring(0, argument.LastIndexOf(" ", StringComparison.Ordinal)).Trim() : argument;
 
@@ -269,97 +268,96 @@ namespace PlantUMLCodeGeneratorGUI
                         if (indexOfAssignment != -1) // Most likely someone is initializing this variable.
                             nextToVisitIndex = indexOfSemiColon;
 
-                        if (nextToVisitIndex == indexOfSemiColon) // This is very likely a variable
+                        if (nextToVisitIndex != indexOfSemiColon) continue; // This is very likely a variable
+
+                        try
                         {
-                            try
+                            var memberContent = methodContent.Trim();
+                            if (memberContent.Length == 0) continue;
+
+                            var ignorablePrefixes = new [] { "friend class ", "typedef ", "using ", "struct ", "class ", "union " /* forward declarations */ };
+                            if (ignorablePrefixes.Any(i => memberContent.StartsWith(i))) continue;
+
+                            // In c++ we can define/declare/initialize variable like this
+                            //
+                            // int a;
+                            // int a, b, c;
+                            // int a = 10, b = 20, c = 30;
+                            // int a = 10, b, c = 30;
+                            // int a{10}, b{20}, c{30};
+                            // int a(10), b(20), c(30);
+                            // int a = {}, b = {}, c = {};
+                            // auto a = 10, b = 20, c = 30;
+                            // int a[] = {1, 2, 3};
+                            // 
+                            // We don't care about the values, just about the type and the variable names
+                            //
+                            // Including the above situation we can have members like this
+                            // int a = 10;
+                            // int a[3];
+                            // int a[] = {1, 2, 3};
+                            // int a[2][2];
+                            // Something a;
+                            // Something a[];
+                            // Something<int> a;
+                            // Something <int> a;
+                            // long long a; // There can even be type modified like this for certain types
+                            // unsigned long long a;
+                            // 
+                            // So we are going to lookup the first , after the >, so that we will get the first
+                            // member in a series of member declarations like "Something a, b, c" or Something<int> a, b, c
+                            // and from out of that we are going to remove the array part if any. So for ex:
+                            // Something<int> a[1][2], b, c => Something<int> a; Then its the item after the last space, which
+                            // is the member name, everything before this is the member type
+
+                            var indexOfLastTemplateEnd = Math.Max(0, memberContent.IndexOf(">", StringComparison.Ordinal));
+                            var indexOfFirstComma = memberContent.IndexOf(",", indexOfLastTemplateEnd, StringComparison.Ordinal);
+                            if (indexOfFirstComma == -1) indexOfFirstComma = memberContent.Length;
+
+                            var firstMember = memberContent.Substring(0, indexOfFirstComma).Trim();
+                            var indexOfFirstEqSign = firstMember.LastIndexOf("=", StringComparison.Ordinal);
+                            if (indexOfFirstEqSign == -1) indexOfFirstEqSign = firstMember.Length;
+
+                            firstMember = firstMember.Substring(0, indexOfFirstEqSign).Trim();
+                            firstMember = RegExs.bitFieldDeclaration.Replace(firstMember, "");
+
+                            var indexOfLastSpace = firstMember.LastIndexOf(" ", StringComparison.Ordinal);
+                            var memberType = firstMember.Substring(0, indexOfLastSpace).Trim();
+                            memberContent = firstMember.Substring(indexOfLastSpace + 1).Trim();
+
+                            var memberContentSegments = memberContent.Split(',');
+                            foreach (var memberContentSegment in memberContentSegments)
                             {
-                                var memberContent = methodContent.Trim();
-                                if (memberContent.Length == 0) continue;
+                                // We are removing the : \d+; in bit fields
+                                var modifiedMemberContentSegment = RegExs.bitFieldDeclaration.Replace(memberContentSegment, "");
+                                var memberName = modifiedMemberContentSegment.Split('=').First().Trim();
+                                var match = RegExs.regexArrayField.Match(memberName);
 
-                                var ignorablePrefixes = new [] { "friend class ", "typedef ", "using ", "struct ", "class ", "union " /* forward declarations */ };
-                                if (ignorablePrefixes.Any(i => memberContent.StartsWith(i))) continue;
+                                // So there can be variables/members in situations like bit fields where we
+                                // don't have to mention variable/member names. We are going to ignore these fields
+                                if (memberName.Length == 0) continue;
 
-                                // In c++ we can define/declare/initialize variable like this
-                                //
-                                // int a;
-                                // int a, b, c;
-                                // int a = 10, b = 20, c = 30;
-                                // int a = 10, b, c = 30;
-                                // int a{10}, b{20}, c{30};
-                                // int a(10), b(20), c(30);
-                                // int a = {}, b = {}, c = {};
-                                // auto a = 10, b = 20, c = 30;
-                                // int a[] = {1, 2, 3};
-                                // 
-                                // We don't care about the values, just about the type and the variable names
-                                //
-                                // Including the above situation we can have members like this
-                                // int a = 10;
-                                // int a[3];
-                                // int a[] = {1, 2, 3};
-                                // int a[2][2];
-                                // Something a;
-                                // Something a[];
-                                // Something<int> a;
-                                // Something <int> a;
-                                // long long a; // There can even be type modified like this for certain types
-                                // unsigned long long a;
-                                // 
-                                // So we are going to lookup the first , after the >, so that we will get the first
-                                // member in a series of member declarations like "Something a, b, c" or Something<int> a, b, c
-                                // and from out of that we are going to remove the array part if any. So for ex:
-                                // Something<int> a[1][2], b, c => Something<int> a; Then its the item after the last space, which
-                                // is the member name, everything before this is the member type
-
-                                var indexOfLastTemplateEnd = Math.Max(0, memberContent.IndexOf(">", StringComparison.Ordinal));
-                                var indexOfFirstComma = memberContent.IndexOf(",", indexOfLastTemplateEnd, StringComparison.Ordinal);
-                                if (indexOfFirstComma == -1) indexOfFirstComma = memberContent.Length;
-
-                                var firstMember = memberContent.Substring(0, indexOfFirstComma).Trim();
-                                var indexOfFirstEqSign = firstMember.LastIndexOf("=", StringComparison.Ordinal);
-                                if (indexOfFirstEqSign == -1) indexOfFirstEqSign = firstMember.Length;
-
-                                firstMember = firstMember.Substring(0, indexOfFirstEqSign).Trim();
-                                firstMember = RegExs.bitFieldDeclaration.Replace(firstMember, "");
-
-                                var indexOfLastSpace = firstMember.LastIndexOf(" ", StringComparison.Ordinal);
-                                var memberType = memberContent.Substring(0, indexOfLastSpace).Trim();
-                                memberContent = firstMember.Substring(indexOfLastSpace + 1).Trim();
-
-                                var memberContentSegments = memberContent.Split(',');
-                                foreach (var memberContentSegment in memberContentSegments)
+                                if (match.Success)
                                 {
-                                    // We are removing the : \d+; in bit fields
-                                    var modifiedMemberContentSegment = RegExs.bitFieldDeclaration.Replace(memberContentSegment, "");
-                                    var memberName = modifiedMemberContentSegment.Split('=').First().Trim();
-                                    var match = RegExs.regexArrayField.Match(memberName);
+                                    memberName = memberName.Substring(0, match.Index).Trim();
 
-                                    // So there can be variables/members in situations like bit fields where we
-                                    // don't have to mention variable/member names. We are going to ignore these fields
-                                    if (memberName.Length == 0) continue;
+                                    var regexArraySegment = new Regex("\\][ \t]*\\[");
+                                    var arrayLengths = regexArraySegment.Replace(match.Value, ";").Replace("[", "").Replace("]", "").Split(';').Select(i => i.Trim()).ToArray();
 
-                                    if (match.Success)
-                                    {
-                                        memberName = memberName.Substring(0, match.Index).Trim();
-
-                                        var regexArraySegment = new Regex("\\][ \t]*\\[");
-                                        var arrayLengths = regexArraySegment.Replace(match.Value, ";").Replace("[", "").Replace("]", "").Split(';').Select(i => i.Trim()).ToArray();
-
-                                        scope.Members.Add(new ArrayMember { Name = memberName, Type = memberType, OwnerClass = this, ArrayLengths = arrayLengths });
-                                    }
-                                    else
-                                    {
-                                        scope.Members.Add(new Member { Name = memberName, Type = memberType, OwnerClass = this });
-                                    }
+                                    scope.Members.Add(new ArrayMember { Name = memberName, Type = memberType, OwnerClass = this, ArrayLengths = arrayLengths });
                                 }
+                                else
+                                {
+                                    scope.Members.Add(new Member { Name = memberName, Type = memberType, OwnerClass = this });
+                                }
+                            }
 
-                                nextToVisitIndex = indexOfSemiColon;
-                                if (indexOfAssignment != -1) visitedIndex = nextToVisitIndex + 1;
-                            }
-                            catch
-                            {
-                                // ignored
-                            }
+                            nextToVisitIndex = indexOfSemiColon;
+                            if (indexOfAssignment != -1) visitedIndex = nextToVisitIndex + 1;
+                        }
+                        catch
+                        {
+                            // ignored
                         }
                     }
                 }
