+++ PlantUMLCodeGeneratorGUI/classes/Namespace.cs
@@ -2,9 +2,8 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Text.RegularExpressions;
-using PlantUMLCodeGeneratorGUI.classes.utils;
 
-namespace PlantUMLCodeGeneratorGUI
+namespace PlantUMLCodeGeneratorGUI.classes
 {
     class Namespace
     {
