+++ PlantUMLCodeGeneratorGUI/classes/Utilities.cs
@@ -3,7 +3,7 @@ using System.Linq;
 using System.Text.RegularExpressions;
 using PlantUMLCodeGeneratorGUI.classes.utils;
 
-namespace PlantUMLCodeGeneratorGUI
+namespace PlantUMLCodeGeneratorGUI.classes
 {
     class Utilities
     {
