+++ PlantUMLCodeGeneratorGUI/frmMain.cs
@@ -3,7 +3,8 @@ using System.Diagnostics;
 using System.IO;
 using System.Linq;
 using System.Windows.Forms;
-using PlantUMLCodeGeneratorGUI.classes.UI;
+using PlantUMLCodeGeneratorGUI.classes;
+using PlantUMLCodeGeneratorGUI.classes.ui;
 
 namespace PlantUMLCodeGeneratorGUI
 {
